refactor enzyme withdraw script into hh task
